refactor: fix the location of exception to be in the service
