+++ chat/src/main/kotlin/net/thechance/chat/api/controller/ChatController.kt
@@ -2,7 +2,6 @@ package net.thechance.chat.api.controller
 
 import net.thechance.chat.api.dto.*
 import net.thechance.chat.service.ChatService
-import net.thechance.chat.service.ContactService
 import net.thechance.chat.service.model.toRequestArgs
 import org.springframework.data.domain.Pageable
 import org.springframework.http.ResponseEntity
@@ -13,7 +12,7 @@ import org.springframework.security.core.annotation.AuthenticationPrincipal
 import org.springframework.stereotype.Controller
 import org.springframework.web.bind.annotation.*
 import java.security.Principal
-import java.util.UUID
+import java.util.*
 
 @RequestMapping("/chat")
 @Controller
