+++ chat/src/main/kotlin/net/thechance/chat/service/exception/ErrorCodes.kt
@@ -1,4 +1,4 @@
-package net.thechance.chat.api.exception
+package net.thechance.chat.service.exception
 
 object ErrorCodes {
     const val INVALID_IMAGE_FORMAT = 1201
